+++ plugins/orchestrator-backend/src/service/DevModeService.ts
@@ -1,5 +1,5 @@
-import { LoggerService } from '@backstage/backend-plugin-api';
-import { Config } from '@backstage/config';
+import type { LoggerService } from '@backstage/backend-plugin-api';
+import type { Config } from '@backstage/config';
 
 import fs from 'fs-extra';