+++ plugins/orchestrator/src/api/api.ts
@@ -6,12 +6,11 @@ import { AxiosResponse } from 'axios';
 import {
   AssessedProcessInstanceDTO,
   ExecuteWorkflowResponseDTO,
-  FilterInfo,
-  PaginationInfoDTO,
+  GetInstancesRequest,
+  InputSchemaResponseDTO,
   ProcessInstanceListResultDTO,
   WorkflowDefinition,
   WorkflowExecutionResponse,
-  WorkflowInputSchemaResponse,
   WorkflowOverviewDTO,
   WorkflowOverviewListResultDTO,
 } from '@janus-idp/backstage-plugin-orchestrator-common';
@@ -39,11 +38,10 @@ export interface OrchestratorApi {
     includeAssessment: boolean,
   ): Promise<AxiosResponse<AssessedProcessInstanceDTO>>;
 
-  getWorkflowDataInputSchema(args: {
-    workflowId: string;
-    instanceId?: string;
-    assessmentInstanceId?: string;
-  }): Promise<WorkflowInputSchemaResponse>;
+  getWorkflowDataInputSchema(
+    workflowId: string,
+    instanceId?: string,
+  ): Promise<AxiosResponse<InputSchemaResponseDTO>>;
 
   getWorkflowOverview(
     workflowId: string,
@@ -53,10 +51,9 @@ export interface OrchestratorApi {
     AxiosResponse<WorkflowOverviewListResultDTO>
   >;
 
-  listInstances(args?: {
-    paginationInfo?: PaginationInfoDTO;
-    filterInfo?: FilterInfo;
-  }): Promise<AxiosResponse<ProcessInstanceListResultDTO>>;
+  listInstances(
+    args?: GetInstancesRequest,
+  ): Promise<AxiosResponse<ProcessInstanceListResultDTO>>;
 }
 
 export const orchestratorApiRef = createApiRef<OrchestratorApi>({