+++ plugins/orchestrator/CHANGELOG.md
@@ -1,5 +1,11 @@
 ### Dependencies
 
+* **@janus-idp/backstage-plugin-orchestrator-common:** upgraded to 1.20.0
+* **@janus-idp/backstage-plugin-orchestrator-form-react:** upgraded to 1.1.6
+* **@janus-idp/cli:** upgraded to 1.15.2
+
+### Dependencies
+
 * **@janus-idp/backstage-plugin-orchestrator-common:** upgraded to 1.19.0
 * **@janus-idp/backstage-plugin-orchestrator-form-react:** upgraded to 1.1.5